+++ CameraPicker/MainPage.xaml
@@ -2,10 +2,13 @@
 <ContentPage xmlns="http://schemas.microsoft.com/dotnet/2021/maui"
              xmlns:x="http://schemas.microsoft.com/winfx/2009/xaml"
              xmlns:imageEditor="clr-namespace:Syncfusion.Maui.ImageEditor;assembly=Syncfusion.Maui.ImageEditor"
+             xmlns:local="clr-namespace:CameraPicker"
              x:Class="CameraPicker.MainPage">
-
     <Grid>
-        <Button Text="Press here to capture an image" Clicked="Button_Clicked" HorizontalOptions="Center" VerticalOptions="Center"/>
+        <Button x:Name="cameraClicked" Text="Press here to capture an image" HorizontalOptions="Center" VerticalOptions="Center"/>
         <imageEditor:SfImageEditor x:Name="imageEditor" IsVisible="false"/>
     </Grid>
+    <ContentPage.Behaviors>
+        <local:CameraPickerBehavior/>
+    </ContentPage.Behaviors>
 </ContentPage>
